+++ packages/labs/lazy/src/lazy-element.ts
@@ -11,6 +11,24 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-export * from './burner-element.js';
-export * from './orchestrator.js';
+import {LitElement} from 'lit-element';
+import {Loader, ElementDependencies, Orchestrator} from './orchestrator.js';
+
 export * from './action.js';
+export * from 'lit-element';
+
+export class LazyElement extends LitElement {
+  static dependencies: Loader;
+
+  static elementDependencies: ElementDependencies;
+
+  static orchestrator: Orchestrator;
+
+  constructor() {
+    super();
+    const ctor = this.constructor as typeof LazyElement;
+    if (ctor.orchestrator !== undefined) {
+      ctor.orchestrator.load(ctor.dependencies);
+    }
+  }
+}